+++ src/workflow/oph_workflow_structs.h
@@ -27,6 +27,8 @@
  * \param markerid Marker ID of the job
  * \param status Status of the task
  * \param ncores Number of ncores to be used in task execution
+ * \param nhosts Number of nhosts to be used in task execution
+ * \param nthreads Number of nthreads to be used in task execution
  * \param arguments_keys Array of explicit parameters keys for the operator
  * \param arguments_values Array of explicit parameters values for the operator
  * \param arguments_num Number of explicit parameters for the operator
@@ -37,6 +39,8 @@ typedef struct _oph_workflow_light_task {
 	int markerid;
 	int status;
 	int ncores;
+	int nhosts;
+	int nthreads;
 	char **arguments_keys;
 	char **arguments_values;
 	int arguments_num;
@@ -72,6 +76,8 @@ typedef struct _oph_workflow_dep {
  * \param operator Name of the operator called by task
  * \param role Permission needed to execute the operator
  * \param ncores Number of ncores to be used in task execution
+ * \param nhosts Number of nhosts to be used in task execution
+ * \param nthreads Number of nthreads to be used in task execution
  * \param arguments_keys Array of explicit parameters keys for the operator
  * \param arguments_values Array of explicit parameters values for the operator
  * \param arguments_num Number of explicit parameters for the operator
@@ -101,6 +107,8 @@ typedef struct _oph_workflow_task {
 	char *operator;
 	int role;
 	int ncores;
+	int nhosts;
+	int nthreads;
 	char **arguments_keys;
 	char **arguments_values;
 	int arguments_num;
@@ -137,6 +145,8 @@ typedef struct _oph_workflow_task {
  * \param sessionid SessionID for the entire workflow
  * \param exec_mode Execution mode for the entire workflow
  * \param ncores Number of cores for the entire workflow
+ * \param nhosts Number of hosts for the entire workflow
+ * \param nthreads Number of hosts for the entire workflow
  * \param cwd CWD for the entire workflow
  * \param cdd CDD for the entire workflow
  * \param command Original command submitted by the user
@@ -164,6 +174,8 @@ typedef struct _oph_workflow {
 	char *sessionid;
 	char *exec_mode;
 	int ncores;
+	int nhosts;
+	int nthreads;
 	char *command;
 	char *cwd;
 	char *cdd;